File rename: model_pipeline renamed to model_engineering.py
